[REFACTOR] Snaps View Header
Adding fading to header on snaps view. It is right now white background but I need to you add same alpha value as feeds section as it is blocking the top part of the image. since it is a box, its on top of the image and when it has full-white background color, its blocking the image behind it. so we need to add alpha to it and change the text colors if its necessary to you

This issue fixed